PycharmProjects/convopro-pvtgpt/requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
pydantic-settings>=2.0.0
# sqlite3 is built into Python — no need to install it!
pytz>=2024.1
